Add fastest server load balancer.
LB strategy that selects the fastest server based on the latest average response time.
It's aimed to handle a multi datacenter setup where some server regularly need longer to response.
It shall also detect short (subsecond) pauses of a server e.g. due to garbage collection or s.th. else.

The latest average response time is determined in the following order (the first found measure is used):
- currently still running requests (if they're lasting longer than previous, already completed requests)
- average response time of the current or the last second
- average response time of the last ten seconds
- total average resonse time

The response time is measured using the given testQuery. A dedicated test query is used, because
queries can have very different performance characteristics, so that it might even be hard for an application
to classify this. With the testQuery you have control what is used to measure response time.

For "normal" / "fast" servers (default: with a resonse time <= the average of all servers, you can override
`determineFastServers`) test queries are run whenever a request comes in, but between test queries at least
the minDelay has to be passed.
This way servers are not hammered with requests when nobody else is using your search, but if there's
high traffic this load balancer get high resolution of monitoring data to detect short pauses as well.

For "slow" servers (default: response time > average) tests are run using the specified maxDelay.

Directly after creation of this LoadBalancer / AsnycSolrClient multiple test queries are run (according
to the specified `initialTestRuns`, by default 10) to have initial stats.