FastestServerLB: key fastServers by collection, get col+testQuery by server.
The collection is used to partition server when classifying "fast"/"slow" servers,
because for different collections response times will be different.
It's somehow similar with the testQuery: it might be different per server, e.g.
some server might only provide a /suggest handler while others provide /select
(which can be specified via the "qt" query param in the test query).