Allow the client to specify the preferred server for a query.
This is useful if there are multiple subsequent, related queries,
where hitting the same server might improve solr cache usage.

For this the new method

```
queryPreferred(q: SolrQuery,
               preferred: Option[SolrServer]): Future[(QueryResponse, SolrServer)]
```
is added.

The RoundRobinLB considers this as expected, the FastestServerLB selects
the preferred server as long as it's one of the fastest servers (after applying
the mapPredictedResponseTime function).