Support several future implementations.
This is built upon pull request inoio/solrs#4 and mainly
adds failure handling (Future.handle/handleWith).
The ScalaFutureFactory now uses a sameThread EC, which is in
line with the latest changes in solrs (sameThread EC was used
lately instead of the global one).

OOTB ScalaFutureFactory and TwitterFutureFactory are provided.

Tests are using the ScalaFutureFactory throughout, only the
TwitterFutureFactorySpec "uses" the twitter factory.

Other changes:
- AsyncSolrClientAware is only implemented by classes that really
  need it (CloudSolrServers and FastestServerLB), and no longer
  by their traits. This was changed, because otherwise the traits
  and all implementing classes would have had to be parameterized
  with the Future type.