Optimize routing: Send updates to leader
Update requests (such as adding docs) are now sent to the shard leader.

This adds the `isLeader` property to the `SolrServer`, which might
require some code changes if it's used by client code, such as filtering
fast servers in `FastestServerLB`. Since this is expected to be rarely
used the major version must not necessarily increased.