Add test collection2, enhance SolrCloudRunner to support symlinks
`solr-home/collection2/conf` is a symlink to `../collection1/conf`
to prevent copying all this stuff in the source tree.

This symlink required some changes:
* copy the solrhome not from the `target/` directory but from the
  source directory (because sbt "copies" the symlink as empty directory)
* no longer use commons-io `FileUtils.copyDirectory` but use our own
  solution, because commons-io does not handle the symlink appropriately
  (and the existing scala better-files does it not better)

The symlink actually is resolved during the copy, i.e. the link
target directory is copied to the destination: necessary because
uploading from the symlink directory failed (in `SolrZkClient.uploadToZK`).

All this stuff is basically a preparation for collection alias tests.